pyproject.toml
[project]
name = "merlin-anki"
version = "0.1.0"
description = "Build Anki flashcard decks for bird identification from allaboutbirds.org"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "genanki>=0.13.1",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
    "ty>=0.0.31",
]
